refactor: api prefix 수정
