+++ api-owner/src/main/java/com/wellmeet/restaurant/RestaurantService.java
@@ -1,9 +1,10 @@
-package com.wellmeet.ownermanage.reservation;
+package com.wellmeet.restaurant;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
-public class ReservationService {
+public class RestaurantService {
+
 }
