+++ api-user/src/main/java/com/wellmeet/favorite/FavoriteController.java
@@ -13,7 +13,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
-@RequestMapping("/api/favorite")
+@RequestMapping("/user/favorite")
 @RestController
 @RequiredArgsConstructor
 public class FavoriteController {
