+++ api-user/src/main/java/com/wellmeet/reservation/ReservationController.java
@@ -19,7 +19,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
-@RequestMapping("/api/reservation")
+@RequestMapping("/user/reservation")
 @RestController
 @RequiredArgsConstructor
 public class ReservationController {
