+++ api-user/src/main/java/com/wellmeet/restaurant/RestaurantController.java
@@ -11,7 +11,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-@RequestMapping("/api/restaurant")
+@RequestMapping("/user/restaurant")
 @RestController
 @RequiredArgsConstructor
 public class RestaurantController {
